requirements.txt
streamlit
PyPDF2
python-dotenv
faiss-cpu
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
sentence-transformers
